+++ setup.py
@@ -36,6 +36,7 @@ setup(
     entry_points={'console_scripts': ['align_on_col = md_utils.align_on_col:main',
                                       'calc_pka = md_utils.calc_pka:main',
                                       'calc_split_avg = md_utils.calc_split_avg:main',
+                                      'col_stats = md_utils.col_stats:main',
                                       'convert_cp2k_forces = md_utils.convert_cp2k_forces:main',
                                       'data2pdb = md_utils.data2pdb:main',
                                       'data_edit = md_utils.data_edit:main',
@@ -52,7 +53,6 @@ setup(
                                       'path_bin = md_utils.path_bin:main',
                                       'pdb_edit = md_utils.pdb_edit:main',
                                       'psf_edit = md_utils.psf_edit:main',
-                                      'per_col_stats = md_utils.per_col_stats:main',
                                       'press_dups = md_utils.press_dups:main',
                                       'proc_plumed_cv = md_utils.proc_plumed_cv:main',
                                       'wham_block = md_utils.wham_block:main',
